requirements.txt
colorama
dearpygui
pyperclip
PyYAML
requests
websocket-client